requirements.txt
absl-py==2.1.0
asttokens==2.4.1
blinker==1.9.0
cachetools==5.3.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
distlib==0.3.7
easyocr==1.7.2
executing==2.0.1
filelock==3.13.1
Flask==3.1.0
Flask-Uploads==0.2.1
fonttools==4.54.1
fsspec==2024.10.0
h5py==3.12.1
imageio==2.36.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.7
lazy_loader==0.4
ldap3==2.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.0
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
optree==0.13.1
packaging==23.2
parso==0.8.4
pd==0.0.4
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.5.0
pyclipper==1.3.0.post6
Pygments==2.18.0
pyparsing==3.2.0
pyproject-api==1.6.1
pytesseract==0.3.10
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
rich==13.9.4
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.5.0
shapely==2.0.6
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tifffile==2024.9.20
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
virtualenv==20.24.6
wcwidth==0.2.13
Werkzeug==3.1.3
